\name{Cat}
\alias{Cat}
\title{Concatenate and print}
\description{
Prints to the command line with carriage return and without newline.
}
\usage{
Cat(message)
}
\arguments{
  \item{message}{}
}
\author{Helmut Strasser}
\seealso{cat}
\examples{
for (i in 1:100) Cat(i)
}
\keyword{misc}
