\name{pExp}
\alias{pExp}
\title{Stochastic exponential}
\description{
The function calculates the stochastic exponential of a
continuous path.
}
\usage{
pExp(p)
}
\arguments{
  \item{p}{ object of \code{class("path") }}
}
\value{
  The function returns the stochastic exponential of a continuous path
  as an object of \code{class("path")}.
}
\author{ Helmut Strasser }
\seealso{\link{as.path}, \link{its2path}, \link{ngrid.default}, \link{pBind}, \link{pClock}, \link{pDelta}, \link{pVariation},
\link{pLag}, \link{pCompress}, \link{pIntegral}, \link{pBracket}, \link{pReduce}, \link{pApply}}
\examples{
}
\keyword{manip}
