\name{statistics}
\alias{Barplot}
\alias{Boxplot}
\alias{Ecdf}
\alias{Hist}
\alias{Normalplot}
\alias{Scatter}
\alias{Shape}
\title{Explorative statistical diagrams}
\description{
Convenience functions for explorative statistical plotting.
}
\usage{
Barplot(x,levels=sort(unique(x)),reference=NULL,...)

Boxplot(x,...)

Hist(x,pdens=NULL,...)

Ecdf(x,pcdf=NULL,...)

Normalplot(x)

Scatter(x)

Shape(x)
}
\arguments{
  \item{x}{}
  \item{pdens}{}
  \item{pcdf}{}
  \item{...}{}
  \item{levels}{}
  \item{reference}{}
}
\author{ Helmut Strasser }
\examples{
}
\keyword{univar}
