\name{Steps}
\alias{Steps}
\alias{Steps.default}
\alias{Steps.path}
\title{Plotting step functions}
\description{
Tools for plotting step functions.
}
\usage{
\method{Steps}{default}(x,t=NULL)

\method{Steps}{path}(x,t=NULL)
}
\arguments{
   \item{x}{}
   \item{t}{}
}
\author{Helmut Strasser}
\examples{

}
\keyword{hplot}
